Make grunt tasks for running sample analysis
This allows someone to take real data they have downloaded from
CloudWatch metrics and feed that into a script that will run a "what if"
analysis - going through the data and feeding it to the decision maker
as if we were at that point in time to see what decision would be made.
Then they can run that analysis through another tool that generates an
HTML page containing a chart showing the consumed usage (from the real
data) and the hypothetical provisioned capacity that would have been
provisioned by the decision maker.

This allows for visualization of changes both to the code and the
user-created configuration of the decision maker.