Add all basic functionality remaining for this library
This makes the library where it can actually be fully used for early
adopters to list DynamoDB tables from a CloudFormation stack, describe
the tables, make capacity decisions, and update the tables.

There are still a lot of rough edges, TODOs, and needed features, but
this will be able to serve as a 0.9.0 release.