Update eslint rules and fix Travis CI
Previously we had to depend on the eslint plugins specifically instead
of relying on the eslint config's dependency on the plugins. This was
only because Travis CI is using an older (2.x) version of npm by
default. Changing our Travis config to upgrade npm fixes this bad
dependency situation and allows us to only rely on the eslint config.