Allow AbsoluteMinimumProvisioned to raise the capacity
Previously the enforcement of the AbsoluteMinimumProvisioned rule only
happened if the capacity was going to be decreasing on this run. That
prohibited a user from raising their configured value for the rule to be
something higher than their current provisioned capacity and having that
config change raise the actual provisioning. We want to support that
usecase, so now the absolute minimum is enforced regardless of whether
the capacity is increasing, decreasing, or staying the same on this run.

For example, if your current capacity is 100, but you want to prepare
for an anticipated surge and you change your AbsoluteMinimumProvisioned
to 400, DCM should immediately raise the provisioning to 400. Now it
will whereas previously it would not. Note that it will raise the next
planned capacity, but other rules might modify or disallow that change.
For instance, the absolute maximum will still be enforced. Also, the
EnforceMaximumIncreaseRule might make DCM stair-step up to your new
absolute minimum, depending on its configuration.